+++ app/src/main/java/com/teamwss/websoso/ui/feedDetail/model/CommentModel.kt
@@ -1,6 +1,6 @@
 package com.teamwss.websoso.ui.feedDetail.model
 
-import com.teamwss.websoso.ui.feed.model.FeedModel.UserModel
+import com.teamwss.websoso.ui.main.feed.model.FeedModel.UserModel
 
 data class CommentModel(
     val user: UserModel,
