+++ app/src/main/java/com/teamwss/websoso/ui/feedDetail/model/FeedDetailModel.kt
@@ -1,6 +1,6 @@
 package com.teamwss.websoso.ui.feedDetail.model
 
-import com.teamwss.websoso.ui.feed.model.FeedModel
+import com.teamwss.websoso.ui.main.feed.model.FeedModel
 
 data class FeedDetailModel(
     val feed: FeedModel,
