+++ app/src/main/java/com/teamwss/websoso/ui/main/feed/adapter/FeedAdapter.kt
@@ -1,14 +1,14 @@
-package com.teamwss.websoso.ui.feed.adapter
+package com.teamwss.websoso.ui.main.feed.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.teamwss.websoso.ui.feed.FeedItemClickListener
-import com.teamwss.websoso.ui.feed.adapter.FeedType.Feed
-import com.teamwss.websoso.ui.feed.adapter.FeedType.ItemType.FEED
-import com.teamwss.websoso.ui.feed.adapter.FeedType.ItemType.LOADING
-import com.teamwss.websoso.ui.feed.adapter.FeedType.Loading
+import com.teamwss.websoso.ui.main.feed.FeedItemClickListener
+import com.teamwss.websoso.ui.main.feed.adapter.FeedType.Feed
+import com.teamwss.websoso.ui.main.feed.adapter.FeedType.ItemType.FEED
+import com.teamwss.websoso.ui.main.feed.adapter.FeedType.ItemType.LOADING
+import com.teamwss.websoso.ui.main.feed.adapter.FeedType.Loading
 
 class FeedAdapter(
     private val feedItemClickListener: FeedItemClickListener,
