+++ app/src/main/java/com/teamwss/websoso/ui/main/feed/model/FeedUiState.kt
@@ -1,4 +1,4 @@
-package com.teamwss.websoso.ui.feed.model
+package com.teamwss.websoso.ui.main.feed.model
 
 data class FeedUiState(
     val loading: Boolean = true,
