+++ app/src/main/res/layout/fragment_feed.xml
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:paddingTop="24dp"
-        tools:context=".ui.feed.FeedFragment">
+        tools:context=".ui.main.feed.FeedFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/cl_feed_none"
