+++ app/src/main/res/layout/item_feed_detail_header.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="feed"
-            type="com.teamwss.websoso.ui.feed.model.FeedModel" />
+            type="com.teamwss.websoso.ui.main.feed.model.FeedModel" />
 
         <variable
             name="commentsCount"
